Moved components from Page to Components in Profile.
